+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.4.2]
 
 - added debouncing to autocompletion in prompt field
+- keeping the processing on errors (e.g. LLM model API overload that keeps retrying)
 
 ## [0.4.1]